# Speedup integer division