--- a/src/number.c
+++ b/src/number.c
@@ -1557,9 +1557,9 @@ doc>
 
 int STk_real_isoddp(SCM n)   /* n MUST be a real */
 {
-  SCM q, r;
+  SCM r = 0;
 
-  integer_division(n, MAKE_INT(2), &q, &r);
+  integer_division(n, MAKE_INT(2), NULL, &r);
   /* We are sure here that r is a real */
   return (fpclassify(REAL_VAL(r)) != FP_ZERO);
 }
@@ -2196,9 +2196,85 @@ DEFINE_PRIMITIVE("abs", abs, subr1, (SCM x))
  * @end lisp
 doc>
  */
+static void int_divide(SCM x, SCM y, SCM *quotient, SCM* remainder, int exact)
+{
+  if (!quotient && !remainder)
+    STk_panic("integer_division called with no quotient nor reminder pointers");
+
+  /* Here, x and y can only be integer or bignum (not real) */
+
+  /* NOTE about the remainder: the GMP accepts 'unsigned integers' and
+     als returns 'unsigned integers' for remainders. We can safely use
+     'long' for these remainders, AND these will always fit a fixnum,
+     because we'll only receive them when we passed fixnums as arguments,
+     and the remainder won't be larger. */
+
+  long quo;              /* temp var for long quotient */
+  long rem = 0;          /* temp var for long remainder */
+  mpz_t q;
+  mpz_t r;
+  int big_q = 0;         /* Is the result quotient a bignum? */
+  int big_r = 0;         /* Is the result remainder a bignum? */
+
+  /* The following sequence of "if"s is written for clarity. Even though
+     it may look like we call INTP more times than necessary, the compiler
+     will later optimize the logic here. */
+
+  /* FIXNUM - FIXNUM */
+  if (INTP(x) && INTP(y)) {
+    long int i1 = INT_VAL(x);
+    long int i2 = INT_VAL(y);
+
+    if (quotient)  quo = i1 / i2;
+    if (remainder) rem = i1 % i2;
+
+  /* FIXNUM - BIGNUM */
+  } else if (INTP(x)) {
+      /* STklos never stores fixnums as if they were bignums, so
+         we're sure that the quotient for INT/BIG = zero. */
+      if (quotient)  quo = 0;
+      if (remainder) rem = INT_VAL(x);
+
+  /* BIGNUM - FIXNUM */
+  } else  if (INTP(y)) {
+      if (quotient) mpz_init(q);
+      /* The GMP only returns unsigned remainders, so we need to keep track of the
+         sign of x. The easiest way to put back the sign is to initialize rem with
+         it, and multiply by whatever the GPM returns.
+         Also, we need labs so GMP will get the expected ulong. */
+      long xsign = mpz_sgn(BIGNUM_VAL(x));
+      if (!quotient) rem = xsign * (long) mpz_tdiv_ui(BIGNUM_VAL(x), labs(INT_VAL(y)));
+      else rem = xsign * (long) mpz_tdiv_q_ui(q, BIGNUM_VAL(x), labs(INT_VAL(y))); /* rem may or may not be used */
+      if (quotient && (INT_VAL(y) < 0)) mpz_neg(q,q); /* Put back the sign of y */
+      big_q = 1;
+
+  /* BIGNUM - BIGNUM */
+  } else {
+    if (quotient)  mpz_init(q);
+    if (remainder) mpz_init(r);
+
+    if (!quotient)        mpz_tdiv_r(r, BIGNUM_VAL(x), BIGNUM_VAL(y));
+    else if (!remainder)  mpz_tdiv_q(q, BIGNUM_VAL(x), BIGNUM_VAL(y));
+    else                  mpz_tdiv_qr(q, r, BIGNUM_VAL(x), BIGNUM_VAL(y));
+
+    big_q = big_r = 1;
+  }
+  /*** END OF CASES ***/
+
+  /* Check exactness and set the result values */
+  if (exact) {
+      if (quotient)  *quotient  = big_q ? bignum2number(q) : long2integer(quo);
+      if (remainder) *remainder = big_r ? bignum2number(r) : long2integer(rem);
+  } else {
+      if (quotient)  *quotient  = big_q ? double2real(bignum2double(q)) : double2real((double) quo);
+      if (remainder) *remainder = big_r ? double2real(bignum2double(r)) : double2real((double) rem);
+  }
+  if (quotient  && big_q) mpz_clear(q);
+  if (remainder && big_r) mpz_clear(r);
+}
+
 static void integer_division(SCM x, SCM y, SCM *quotient, SCM* remainder)
 {
-  mpz_t q, r;
   int exact = 1;
 
   if (!INTP(x) && !BIGNUMP(x) && !REALP(x)) error_bad_number(x);
@@ -2207,69 +2283,32 @@ static void integer_division(SCM x, SCM y, SCM *quotient, SCM* remainder)
 
   if (REALP(x)) { x = real2integer(x); exact = 0; }
   if (REALP(y)) { y = real2integer(y); exact = 0; }
-
-  /* Here, x and y can only be integer or bignum (not real) */
-  if (INTP(x)) {
-    if (INTP(y)) {
-      long int i1 = INT_VAL(x);
-      long int i2 = INT_VAL(y);
-
-      if (exact) {
-        *quotient  = MAKE_INT(i1 / i2);
-        *remainder = MAKE_INT(i1 % i2);
-      } else {
-        /* Useless casts to long are here for clang-tidy */
-        *quotient  = double2real((double) ((long) (i1 / i2)));
-        *remainder = double2real((double) ((long) (i1 % i2)));
-      }
-      return;
-    }
-    else
-      x = long2scheme_bignum(INT_VAL(x));
-  } else {
-    /* x is a bignum */
-    if (INTP(y))
-      y = long2scheme_bignum(INT_VAL(y));
-  }
-
-  /* Here x and y are both bignum */
-  mpz_init(q); mpz_init(r);
-  mpz_tdiv_qr(q,r,BIGNUM_VAL(x),BIGNUM_VAL(y));
-  if (exact) {
-    *quotient  = bignum2number(q);
-    *remainder = bignum2number(r);
-  } else {
-    /*     *quotient  = double2real(mpz_get_d(q)); */
-    /*     *remainder = double2real(mpz_get_d(r)); */
-    *quotient  = double2real(bignum2double(q));
-    *remainder = double2real(bignum2double(r));
-  }
-  mpz_clear(q); mpz_clear(r);                          /* //FIXME: TESTER */
+  int_divide(x,y,quotient,remainder,exact);
 }
 
 DEFINE_PRIMITIVE("quotient", quotient, subr2, (SCM n1, SCM n2))
 {
-  SCM q, r;
+  SCM q = 0;
 
-  integer_division(n1, n2, &q, &r);
+  integer_division(n1, n2, &q, NULL);
   return q;
 }
 
 
 DEFINE_PRIMITIVE("remainder", remainder, subr2, (SCM n1, SCM n2))
 {
-  SCM q, r;
+  SCM r = 0;
 
-  integer_division(n1, n2, &q, &r);
+  integer_division(n1, n2, NULL, &r);
   return r;
 }
 
 
 DEFINE_PRIMITIVE("modulo", modulo, subr2, (SCM n1, SCM n2))
 {
-  SCM q, r;
+  SCM r = 0;
 
-  integer_division(n1, n2, &q, &r);
+  integer_division(n1, n2, NULL, &r);
   if (negativep(n1) != negativep(n2) && !zerop(r))
      /*kerch@parc.xerox.com*/
     r = add2(r, n2);