Both the GMP and the mini-GMP have functions to extract only the quotient and only the remainder from a division. We change `integer_division` so that if either of `quotient` or `remainder` is `NULL`, then it is not used, and a faster GMP function is called for bignums.

We also do the same for fixnums and doubles, for consistency and simplicity (so as not to treat bignums in a separate function), although there doesn't seem to be much of a speedup for fixnums.

TIMINGS
=======

Ten runs of each, the time reported is the average.

Bignums
-------

```scheme
(let ((big (+ (expt 2 70) 3))
      (big2 (+ (expt 3 62) 99877))
      (a 1))
  (time
    (repeat 10_000_000
      (set! a (quotient big2 big))))
  a)

current code: 3984.07 ms
this patch:   1781.56 ms
```

```scheme
(let ((big (+ (expt 2 70) 3))
      (big2 (+ (expt 3 62) 99877))
      (a 1))
  (time
    (repeat 10_000_000
      (set! a (remainder big2 big))))
  a)

current code: 3958.14 ms
this patch:   3360.73 ms
```

```scheme
(let ((big (+ (expt 2 70) 3))
      (big2 (+ (expt 3 62) 99877))
      (a 1))
  (time
    (repeat 10_000_000
      (set! a (modulo big2 big))))
  a)

current code: 4026.56 ms
this patch:   3515.91 ms
```

Flonums
-------

```scheme
(let ((a 1))
  (time
    (repeat 20_000_000
      (set! a (quotient 10928.0 987.0))))
  a)

current code: 4010.85 ms
this patch:   2887.87 ms
```

```scheme
(let ((a 1))
  (time
    (repeat 20_000_000
      (set! a (remainder 10928.0 987.0))))
  a)

current code: 3940.73 ms
this patch:   2890.27 ms
```

```scheme
(let ((a 1))
  (time
    (repeat 20_000_000
      (set! a (modulo 10928.0 987.0))))
  a)

current code: 4073.73 ms
this patch:   2962.37 ms
```

Fixnums
-------

```scheme
(let ((a 1))
  (time
    (repeat 20_000_000
      (set! a (quotient 10928 987))))
  a)

current code: 1418 ms
this patch:   1420 ms
```

```scheme
(let ((a 1))
  (time
    (repeat 20_000_000
      (set! a (remainder 10928 987))))
  a)

current code: 1455.19 ms
this patch:   1450.22 ms
```

```scheme
(let ((a 1))
  (time
    (repeat 20_000_000
      (set! a (modulo 10928 987))))
  a)

current code: 1513.41 ms
this patch:   1526.12 ms
```

`Odd?` and `even?`for flonums
----------------

Interestingly, this one get really faster:

```scheme
(let ((a 1)
      (k (+ 1.0 (expt 3 63))))
  (time
    (repeat 10_000_000
      (set! a (odd? k))))
  a)

current code: 12033.47 ms
this patch:    6485.03 ms
```