@jpellegrini 2023-04-15
I think this can be enhanced:
* `int_quotient` could be replaced by `integer_division` since `integer_division` now can be asked to only return the quotient
* the code for `integer_division` can be a bit more clear

I'll change this PR soon