@jpellegrini 2023-04-15
Ok - I have made a better version of it. It's clearer.

I have also simplified `int_quotient` so it uses one less `mpz_t` variable, and calls `mpz_clear` on exit.
This one os ready.

By the way - the timings in the commit message have been updated - the new version seems better than the first one I had submitted in the PR