@jpellegrini 2023-04-16
And the really nice speedup happens when we divide a fixnum by a bignum:
```scheme
(let ((big (+ (expt 2 70) 3))
      (fix 10010192)
      (a 1))
  (time
    (repeat 10_000_000
      (set! a (remainder fix big))))
  a)

current code: 3258.373 ms
this patch:    762.4 ms
```

```scheme
(let ((big (+ (expt 2 70) 3))
      (fix 10010192)
      (a 1))
  (time
    (repeat 10_000_000
      (set! a (quotient fix big))))
  a)
current code: 3306.894 ms
this patch:    776.137 ms
```