@jpellegrini 2023-04-16
Ok, so I included some more commits related to integer division, including a considerable speedup to `round` (the idea is to do integer division, then just check if  `abs(remainder) / 2` is past the midpoint of all possible remainders.

All tests pass. I'm not sure there is more to include in this PR -- I believe it's ready.