@egallesio 2023-07-02
Hello @jpellegrini,

I have merged this PR. The enhancement is really impressive :+1: . For fixnum operations, you had similar performance with and without the patch. On my laptop, (an old i5), these operations were 10% slower with the patch. I have finally added a fast path in `integer-division` when arguments are both fixnums (without any conversion) and the code is now faster than before also in this case (which is probably the more frequent). 
Thanks a lot for this contribution.